pyproject.toml
[tool.poetry]
name = "telegram-banking"
version = "0.1.0"
description = ""
authors = ["dakohhh <wiizzydreadmill@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
aiogram = "^3.20.0.post0"
python-dotenv = "^1.1.0"
pydantic-settings = "^2.9.1"
alembic = "^1.15.2"
certifi = "^2025.4.26"
sqlmodel = "^0.0.24"
asyncpg = "^0.30.0"
psycopg2 = "^2.9.10"
pendulum = "^3.1.0"
httpx = "^0.28.1"
openai-agents = "^0.0.14"


[tool.poetry.group.dev.dependencies]
watchdog = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
